Convert client checkout logic to a reusable hook.
